requirements.txt
requests
boto3
urllib3==1.26.6
finnhub-python